system/index/result_set: lookups on the index return a live updating result set
queries for values at a key or KeySet can be made against an index by
using the +lookup+ function.  a ResultSet will be the return of the
lookup.  ResultSet automatically subscribe to the didChange/willChange
callbacks on the index to watch all changes in the index.  when a
change occurs that affects one of its key set values then the
ResultSet updates its IndexSet by requerying the index for an updated
version.